requirements.txt
blinker==1.8.2
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
Flask==3.0.3
Flask-Email==1.4.4
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gmpy2==2.2.1
google-ai-generativelanguage==0.1.0
google-api-core==2.28.1
google-auth==2.43.0
google-generativeai==0.1.0rc1
googleapis-common-protos==1.72.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.62.3
gunicorn==21.2.0
idna==3.11
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
packaging==25.0
pillow==10.4.0
proto-plus==1.26.1
protobuf==4.25.8
psycopg==3.2.12
psycopg-binary==3.2.12
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.21.0
python-dotenv==1.0.1
requests==2.32.4
rsa==4.9.1
SQLAlchemy==2.0.44
typing_extensions
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
gunicorn
Flask-Migrate
flask_bcrypt